requirements.txt
click==8.0.3
cycler==0.11.0
Flask==2.0.2
fonttools==4.28.2
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
numpy==1.21.4
opencv-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
Werkzeug==2.0.2
